requirements.txt
appdirs==1.4.4
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pipenv==2020.6.2
pycparser==2.20
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
urllib3==1.25.9
URLObject==2.4.3
virtualenv==20.0.26
virtualenv-clone==0.5.4
Werkzeug==1.0.1
WTForms==2.3.1
